+++ Roguelike/AreaPath.h
@@ -9,9 +9,9 @@ public:
 	std::vector<Tile*> blockingTiles;
 	std::vector<Tile*> overlayTiles;
 	Point2D destination;
-	float weightToUseExistingPath;
+	double weightToUseExistingPath;
 	int width;
-	float diagonalCost;
+	double diagonalCost;
 
 	std::shared_ptr<TCODPath> pathMap;
 	class PathCostCallback : public ITCODPathCallback{ public: float getWalkCost(int xFrom, int yFrom, int xTo, int yTo, void *userData) const; };
@@ -19,6 +19,6 @@ public:
 
 	void build(Point2D &from, Point2D &to);
 
-	AreaPath(Tile &pathTile, Area &area, std::vector<Tile*> blockingTiles, std::vector<Tile*> overlayTiles, float weightToUseExistingPath, int width = 1, float diagonalCost = 1.41f) :
+	AreaPath(Tile &pathTile, Area &area, std::vector<Tile*> blockingTiles, std::vector<Tile*> overlayTiles, double weightToUseExistingPath, int width = 1, double diagonalCost = 1.41f) :
 		pathTile(pathTile), area(area), blockingTiles(blockingTiles), overlayTiles(overlayTiles), weightToUseExistingPath(weightToUseExistingPath), width(width), diagonalCost(diagonalCost){};
 };
