+++ Roguelike/LogFrame.h
@@ -16,7 +16,7 @@ public:
 	void GuiFrame::render();
 	bool handleKey(TCOD_key_t key);
 
-	LogFrame(char controlKey, bool open, float alphaFg, float alphaBg) :
+	LogFrame(char controlKey, bool open, double alphaFg, double alphaBg) :
 		GuiFrame(controlKey, open, alphaFg, alphaBg){};
 };
